[Triggered] Cog model restructure
- Restructure `Triggered` according to SFUAnime/Ren#608. No functional
  changes.
- Remove redundant imports.
- Remove redundant constant `AVATAR_URL`.
- Sort imports in this order: standard library imports, third-party
  imports, upstream (`redbot`) imports, and lastly, our code.
- Rework enum `Modes` by making it an `Enum`-derived class.
